Moving SearchBox into react-internal package
