--- apps/perf-test/src/scenarios/SearchBoxNext.tsx
@@ -1,6 +1,0 @@
-import * as React from 'react';
-import { SearchBox } from '@fluentui/react-next';
-
-const Scenario = () => <SearchBox placeholder="Search" />;
-
-export default Scenario;
