--- packages/react-examples/src/react-next/SearchBox/SearchBox.FullSize.Example.tsx
@@ -1,19 +1,0 @@
-import * as React from 'react';
-import { SearchBox } from '@fluentui/react-next/lib/SearchBox';
-import { Stack, IStackTokens } from '@fluentui/react-next/lib/Stack';
-
-const stackTokens: Partial<IStackTokens> = { childrenGap: 20 };
-
-/* eslint-disable react/jsx-no-bind */
-export const SearchBoxFullSizeExample = () => {
-  return (
-    <Stack tokens={stackTokens}>
-      <SearchBox placeholder="Search" onSearch={newValue => console.log('value is ' + newValue)} />
-      <SearchBox
-        placeholder="Search with no animation"
-        onSearch={newValue => console.log('value is ' + newValue)}
-        disableAnimation
-      />
-    </Stack>
-  );
-};
