--- packages/react-examples/src/react-next/SearchBox/SearchBox.Underlined.Example.tsx
@@ -1,4 +1,0 @@
-import * as React from 'react';
-import { SearchBox } from '@fluentui/react-next/lib/SearchBox';
-
-export const SearchBoxUnderlinedExample = () => <SearchBox placeholder="Search" underlined={true} />;
