--- packages/react-examples/src/react-next/SearchBox/docs/SearchBoxBestPractices.md
@@ -1,12 +1,0 @@
-### Layout
-
-- Don't build a custom search control based on the default text box or any other control.
-- Use a search box without a parent container when it's not restricted to a certain width to accommodate other content. This search box will span the entire width of the space it's in.
-
-### Content
-
-- Use placeholder text in the search box to describe what people can search for. For example, "Search", "Search files", or "Search contacts list".
-- Although search entry points tend to be similarly visualized, they can provide access to results that range from broad to narrow. By effectively communicating the scope of a search, you can ensure that people's expectations are met by the capabilities of the search you're performing, which will reduce the possibility of frustration. The search entry point should be placed near the content being searched.<br/><br/>Some common search scopes include:
-  - **Global**: Search across multiple sources of cloud and local content. Varied results include URLs, documents, media, actions, apps, and more.
-  - **Web**: Search a web index. Results include pages, entities, and answers.
-  - **My stuff**: Search across devices, cloud, social graphs, and more. Results are varied but are constrained by the connection to user accounts.
