--- packages/react-examples/src/react-next/SearchBox/docs/SearchBoxOverview.md
@@ -1,1 +1,0 @@
-A search box (`SearchBox`) provides an input field for searching content within a site or app to find specific items.
