--- packages/react-next/src/components/SearchBox/SearchBox.tsx
@@ -1,11 +1,0 @@
-import * as React from 'react';
-import { styled } from '../../Utilities';
-import { SearchBoxBase } from './SearchBox.base';
-import { ISearchBoxProps, ISearchBoxStyleProps, ISearchBoxStyles } from './SearchBox.types';
-import { getStyles } from './SearchBox.styles';
-
-export const SearchBox: React.FunctionComponent<ISearchBoxProps> = styled<
-  ISearchBoxProps,
-  ISearchBoxStyleProps,
-  ISearchBoxStyles
->(SearchBoxBase, getStyles, undefined, { scope: 'SearchBox' });
