--- packages/react-next/src/components/SearchBox/index.ts
@@ -1,3 +1,0 @@
-export * from './SearchBox';
-export * from './SearchBox.base';
-export * from './SearchBox.types';
